--- cucTestProject/src/test/java/cucTestProject/BaseClass.java
@@ -1,20 +1,0 @@
-package cucTestProject;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-
-import io.github.bonigarcia.wdm.WebDriverManager;
-
-public class BaseClass {
-	static WebDriver driver;
-	
-	public static void BasePage() {
-		
-		WebDriverManager.chromedriver().setup();
-		driver = new ChromeDriver();
-		
-	}
-	
-	
-
-}
