+++ cucTestProject/src/test/java/cucTestProject/RunnerClass.java
@@ -7,8 +7,8 @@ import io.cucumber.junit.CucumberOptions;
 
 @RunWith(Cucumber.class)
 @CucumberOptions(
-		features = {"src\\test\\java\\login.feature"},
-		glue={"cucTestProject.CucClass"}
+		features = {"C:\\Users\\dhineshkumar_selvam\\git\\incubationprogram\\cucTestProject\\src\\test\\java\\featurefile\\login.feature"},
+		glue={""}
 		
 		)
 public class RunnerClass {
